+++ iOS_SDK/OneSignalSDK/OneSignalInAppMessages/Controller/OSTriggerController.h
