+++ iOS_SDK/OneSignalSDK/OneSignalInAppMessages/Model/OSInAppMessageClickResult.m
