+++ iOS_SDK/OneSignalSDK/OneSignalInAppMessages/Model/OSInAppMessageInternal.h
