+++ iOS_SDK/OneSignalSDK/OneSignalInAppMessages/Model/OSInAppMessageInternal.m
