+++ iOS_SDK/OneSignalSDK/OneSignalInAppMessages/Model/OSInAppMessagePrompt.h
