+++ iOS_SDK/OneSignalSDK/OneSignalInAppMessages/Model/OSTrigger.m
